Move log data to its own file
