+++ packages/devtools_app/lib/src/screens/logging/logging_screen_v2/logging_screen_v2.dart
@@ -13,6 +13,7 @@ import '../../../shared/analytics/constants.dart' as gac;
 import '../../../shared/screen.dart';
 import '../../../shared/utils.dart';
 import 'logging_controller_v2.dart';
+import 'log_data.dart';
 import 'logging_table_v2.dart';
 
 /// Presents logs from the connected app.
