+++ packages/devtools_app/lib/src/screens/logging/logging_screen_v2/logging_table_row.dart
@@ -13,6 +13,7 @@ import '../../../shared/globals.dart';
 import '../../../shared/ui/icons.dart';
 import '../../../shared/ui/utils.dart';
 import 'logging_controller_v2.dart';
+import 'log_data.dart';
 
 final loggingTableTimeFormat = DateFormat('HH:mm:ss.SSS');
