+++ packages/devtools_app/lib/src/screens/logging/logging_screen_v2/logging_table_v2.dart
@@ -15,6 +15,7 @@ import '../../../shared/globals.dart';
 import '../../../shared/ui/filter.dart';
 import '../../../shared/utils.dart';
 import 'logging_controller_v2.dart';
+import 'log_data.dart';
 import 'logging_model.dart';
 import 'logging_table_row.dart';
